separate User model to shared folder
