+++ src/api/consts.ts
@@ -1,1 +1,2 @@
 export const baseUrl = 'https://ya-praktikum.tech/api/v2';
+export const localBaseUrl = 'https://local.ya-praktikum.tech/api/v1'
