+++ src/server/db/models/ForumThread/types.ts
@@ -1,7 +1,5 @@
 import { Model, Optional } from 'sequelize';
 import { CommentAttributes, CommentCreationAttributes } from '../Comment';
-// import { CommentAttributes, CommentCreationAttributes } from '../Comment';
-// import { UserAttributes, UserCreationAttributes } from '../User';
 
 export interface ForumThreadAttributes {
   id: number;
