+++ src/server/db/models/User/User.ts
@@ -1,14 +1,21 @@
 import { ModelDefined, DataTypes } from 'sequelize';
 import { sequelize } from '../../sequelize';
 
-import { UserAttributes, UserCreationAttributes } from './types';
+import { UserCreationAttributes } from './types';
 import { SiteTheme } from '../SiteTheme';
+import { UserAttributes } from 'shared/types/UserAttributes';
 
 const User: ModelDefined<UserAttributes, UserCreationAttributes> = sequelize.define(
   'User',
   {
     // id: { type: DataTypes.INTEGER, autoIncrement: true, primaryKey: true },
-    id: { type: DataTypes.INTEGER, allowNull: false, primaryKey:true, autoIncrement:false, unique: true },
+    id: {
+      type: DataTypes.INTEGER,
+      allowNull: false,
+      primaryKey: true,
+      autoIncrement: false,
+      unique: true,
+    },
     name: { type: DataTypes.STRING(255), allowNull: false },
     avatar: { type: DataTypes.TEXT, allowNull: true },
   },
