+++ src/server/db/models/User/index.ts
@@ -1,2 +1,2 @@
 export { User } from './User';
-export { UserAttributes, UserCreationAttributes } from './types';
+export { UserCreationAttributes } from './types';
