+++ src/shared/types/UpdateModelAttributes.ts
@@ -1,0 +1,4 @@
+export type UpdateModelAttributes = {
+  createdAt: string;
+  updatedAt: string;
+};
