+++ src/shared/types/UserAttributes.ts
@@ -1,0 +1,7 @@
+export interface UserAttributes {
+  id: number;
+  name: string;
+  avatar?: string | null;
+
+  siteThemeId?: number | null;
+}
